requirements.txt
numpy~=1.18.5
attrs~=19.3.0
scipy~=1.4.1
matplotlib~=3.1.2
